Seems work for only voltage source and resistor
The bug is due to the 0th row and column, which made the matrix into
indefinite. Hence, I build a new matrix that removes 0th row/column.
However, this is a little nasty in code and not consistent, so I decided
to just keep the 0th row and column, but sets [0][0] = 1, and [0][i] =
[j][0] = 0, that should works. This will create a new branch.