Fixed a bug in update_node_voltages
Caused by dampen. When doing ac analysis, all the node voltages
are set to 0. Hence when the node voltage are being updated,
their values will changed to actual value, but it will be dampened.
So a boolean parameter is added to update_node_voltages to control
the switch of dampening